try this
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@34321 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: 92e3b1dc55495590a0a74dea62cc7a31cfb69645

+++ src/linux/universal/linux-4.14/include/net/netfilter/nf_conntrack_ecache.h
@@ -15,7 +15,7 @@
 
 struct nf_conntrack_ecache {
 	unsigned long cache;		/* bitops want long */
-	unsigned long missed;			/* missed events */
+	u16 missed;			/* missed events */
 	u16 ctmask;			/* bitmask of ct events to be delivered */
 	u16 expmask;			/* bitmask of expect events to be delivered */
 	u32 portid;			/* netlink portid of destroyer */
